Address issues brought up in review
